requirements.txt
google-adk
google-generativeai
fastapi
uvicorn
chromadb
python-dotenv
requests
pydantic
streamlit
